requirements.txt
streamlit
requests
google-generativeai
jinja2
pdflatex
